add 2 new tests and rename monkey patch function for future extension
